Create AprilSNOTELanom.py
Analysis uses SNOTEL data that will soon be published in v2 of https://zenodo.org/records/10287093 . These are the 'raw' SNOTEL data, not eh BCQC data.

Calculates the mean SWE for 1991-2020 on 1 April. Then calculates the 2024 1 April anomaly relative to 1991-2020.

For each site, the script identifies instances of 0swe that are included in the calculation and calculates the number of years with observations. 

For visualization and analysis, I only mapped sites with observations in all 30yrs.